requirements.txt
yfinance==0.2.51
pandas==2.2.3
requests==2.32.3
pyyaml==6.0.2
